Add dummy audio sample and playback support
Add a 1s 'chopping' dummy audio header (kitchen/src/dummy_audio.h) plus a Python generator (kitchen/dummy.py) and several datatest .wav files to enable testing without I2S hardware. Introduce USE_DUMMY_AUDIO in config.h and update audio code: include dummy_audio.h under the flag, switch the audio buffer and get_audio() to int16_t, add a branch in record_audio() to copy scaled dummy samples (with a simulation delay), and keep normal I2S code under #ifndef. Also add build_error.log, a small platformio.ini whitespace tweak, and update other sources (features.cpp, inference.cpp) as part of this change. Purpose: allow offline/local model testing using an embedded sample and reduce RAM usage by using 16-bit audio samples.

+++ kitchen/src/audio.h
@@ -1,5 +1,6 @@
 #pragma once
+#include <stdint.h>
 
 void audio_init();
 void record_audio();
-float* get_audio();
+int16_t* get_audio();
